+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Remove `{@inheritDoc}` as it only explicitly indicates that the default behaviour at the expense of significant visual clutter.
+* Ensure `AbstractContext` does not NPE when methods are invoked on it after it has been closed.
 
 ## 1.4: